Added a class to manage Kills Remaining Text
Added a new class to manage the kills remaining text on the top right corner of level 1.
